+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 - Changed leak detection to default to `LeakDetectionLevel.DISABLED` instead of `LeakDetectionLevel.SIMPLE` so that leak detection can be opted in.
 - Changed `ResourceLeakDetector` to not instantiate a `Cleaner` when one is not needed (leak detection is disabled).
+- Added `Fields.IgnoreAccess.getStatic(Class, Field)` method to return the given static field value.
 - Added `Fields.Safe.get` that returns `null` instead of throwing exceptions when field is missing or not accessible.
 - Added `Fields.Safe.getStatic` that returns `null` instead of throwing exceptions when field is missing or not accessible.
+- Added `MemberAccessor.isAccessible` to check if a member is accessible.
+- Added `MemberAccessor.setAccessible` to set the access for a member.
 - Upgraded `pitest-maven` to `1.23.1`.
 - Upgraded parent to `4.0.6`.